common/pyproject.toml
[project]
name = "common"
version = "1.0.0"
description = "Common code for Tradesignals pipeline applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.10",
    "apscheduler>=3.10.4",
    "jsonpath-ng>=1.7.0",
    "python-dotenv>=1.0.1",
    "python-json-logger>=2.0.7",
    "quixstreams>=3.1.1",
    "websocket-client>=1.8.0",
]
